api/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pytesseract==0.3.10
Pillow==10.2.0
python-multipart==0.0.6
pydantic==1.10.13
python-dotenv==1.0.0
